+++ .eslintrc.json
@@ -6,11 +6,7 @@
     "jest": true,
     "node": true
   },
-  "plugins": ["@typescript-eslint/eslint-plugin",
-    "eslint-plugin-tsdoc","import"],
-  "parser": "@typescript-eslint/parser",
   "parserOptions": {
-    "moduleResolution": "node",
     "sourceType": "module",
     "ecmaVersion": 2021,
     "ecmaFeatures": {
@@ -18,42 +14,19 @@
     }
   },
   "settings": {
-    "noInlineConfig": true,
-    "import/parsers": {
-      "@typescript-eslint/parser": [".ts", ".tsx"]
-    },
-    "import/resolver": {
-      "typescript": {
-        "alwaysTryTypes": true, // always try to resolve types under `<root>@types` directory even it doesn't contain any source code, like `@types/unist`
-        "project": "./tsconfig.json"
-      }
-    }
-    },
+    "noInlineConfig": true
+  },
   "extends": [
     "eslint:recommended",
-    "plugin:import/recommended",
     "plugin:node/recommended",
     "airbnb-base",
-    "prettier",
-    "plugin:@typescript-eslint/recommended"
+    "prettier"
   ],
   "rules": {
     "import/extensions": "off",
     "linebreak-style": "off",
     "node/no-unsupported-features/es-syntax": "off",
     "no-underscore-dangle": "off",
-    "import/prefer-default-export": "off",
-    "node/no-extraneous-import": ["error", {
-      "allowModules": ["express"]
-
-  }],
-  "node/no-missing-import": ["error", {
-    "allowModules": [],
-    "resolvePaths": ["src"],
-    "tryExtensions": [".js", ".json", ".node",".ts"]
-}],
-  "import/no-unresolved": "error",
-  "@typescript-eslint/no-explicit-any": "error",
-  "tsdoc/syntax": "warn"
+    "import/prefer-default-export": "off"
   }
 }
