+++ package.json
@@ -7,10 +7,9 @@
   },
   "type": "commonjs",
   "description": "REST service starter for RS School",
-  "main": "dist/server.js",
+  "main": "src/server.js",
   "scripts": {
-    "start": "tsc && node .",
-    "dev": "nodemon --watch './**/*.ts' --exec ts-node src/server.ts",
+    "start": "nodemon src/server.js",
     "lint": "eslint --ext .js,.ts --fix src/",
     "test": "cross-env DEBUG=rs:* jest --testMatch \"<rootDir>/test/e2e/test/*.test.js\"",
     "test:auth": "cross-env DEBUG=rs:* TEST_MODE=auth jest --noStackTrace"
@@ -43,7 +42,6 @@
     ]
   },
   "dependencies": {
-    "@microsoft/tsdoc": "^0.13.2",
     "cross-env": "7.0.3",
     "dotenv": "10.0.0",
     "fastify": "^3.24.1",
@@ -52,28 +50,20 @@
     "yamljs": "0.3.0"
   },
   "devDependencies": {
-    "@types/node": "^16.11.12",
-    "@types/uuid": "^8.3.3",
-    "@typescript-eslint/eslint-plugin": "^5.6.0",
-    "@typescript-eslint/parser": "^5.6.0",
     "chai": "4.3.4",
     "debug": "4.3.2",
     "dirty-chai": "2.0.1",
     "eslint": "8.2.0",
     "eslint-config-airbnb-base": "15.0.0",
     "eslint-config-prettier": "8.3.0",
-    "eslint-import-resolver-typescript": "^2.5.0",
-    "eslint-plugin-import": "^2.25.3",
+    "eslint-plugin-import": "2.25.3",
     "eslint-plugin-node": "11.1.0",
     "eslint-plugin-prettier": "4.0.0",
-    "eslint-plugin-tsdoc": "^0.2.14",
     "husky": "^4.3.8",
     "jest": "27.3.1",
     "lint-staged": "12.0.3",
     "nodemon": "2.0.15",
     "prettier": "2.4.1",
-    "supertest": "6.1.6",
-    "ts-node": "^10.4.0",
-    "typescript": "^4.5.3"
+    "supertest": "6.1.6"
   }
 }
