--- src/interfaces/IConfig.ts
@@ -1,7 +1,0 @@
-export default interface IConfig {
-  PORT: number,
-  NODE_ENV: string ,
-  MONGO_CONNECTION_STRING: string,
-  JWT_SECRET_KEY: string ,
-  AUTH_MODE: boolean,
-}
