+++ src/resources/users/user.handler.js
@@ -1,0 +1,45 @@
+const {getAllUsers, getUserById, createUser, updateUserById, deleteUserById} = require('./user.memory.repository');
+
+const getAllUsersHandler = async(req, reply) => {
+  const users = await getAllUsers();
+  reply.code(200).send(users);
+}
+
+const getUserByIdHandler = async(req, reply) => {
+  const { id } = req.params;
+  const user = await getUserById(id);
+  if(!user) {
+    reply.code(404).send();
+  }
+  reply.code(200).send(user);
+}
+
+const createUserHandler = async(req, reply) => {
+  const user = req.body;
+  const newUser = await createUser(user);
+  if(!newUser) {
+    reply.code(400).send();
+  }
+  reply.code(201).send(newUser);
+}
+
+const updateUserByIdHandler = async(req, reply) => {
+  const user = req.body;
+  const { id } = req.params;
+  const updatedUser = await updateUserById(user,id);
+  if(!updatedUser) {
+    reply.code(400).send();
+  }
+  reply.code(200).send(updatedUser);
+}
+
+const deleteUserByIdHandler = async(req, reply) => {
+  const { id } = req.params;
+  const isDeleted = await deleteUserById(id);
+  if(!isDeleted) {
+    reply.code(401).send();
+  }
+  reply.code(204).send();
+}
+
+module.exports = { getAllUsersHandler,getUserByIdHandler,createUserHandler,updateUserByIdHandler,deleteUserByIdHandler };
