--- src/types/Board.type.ts
@@ -1,7 +1,0 @@
-import { Column } from "./Column.type";
-
-export type Board = {
-  id?: string,
-  title: string,
-  columns: Column[],
-}
