--- src/types/Column.type.ts
@@ -1,4 +1,0 @@
-export type Column = {
-  title: string,
-  order: number | null
-}
