--- src/types/User.type.ts
@@ -1,6 +1,0 @@
-export type User = {
-  id?: string,
-  name: string,
-  login: string,
-  password?: string
-}
